# openapi-spec.json
{
  "openapi": "3.1.1",
  "paths": {
...
    "/api/v1/app/widgets": {
      "get": {
        "operationId": "WidgetsController_getWidgets",
        "parameters": [
          {
            "name": "category",
            "required": false,
            "in": "query",
            "description": "category",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "limit",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "exclude",
            "required": false,
            "in": "query",
            "description": "exclude tags",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "description": "type",
            "schema": {
              "enum": [
                "iframe",
                "system"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WidgetSummaryDto"
                  }
                }
              }
            }
          }
        },
        "summary": "get all widgets",
        "tags": [
          "Widgets"
        ]
      }
    },
    "/api/v1/app/widgets/my": {
      "get": {
        "operationId": "WidgetsController_getMyWidgets",
        "parameters": [
          {
            "name": "category",
            "required": false,
            "in": "query",
            "description": "category",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "limit",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "exclude",
            "required": false,
            "in": "query",
            "description": "exclude tags",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "description": "type",
            "schema": {
              "enum": [
                "iframe",
                "system"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WidgetSummaryDto"
                  }
                }
              }
            }
          }
        },
        "summary": "get all my widgets",
        "tags": [
          "Widgets"
        ]
      }
    },
    "/api/v1/app/widgets/getConfigs": {
      "get": {
        "operationId": "WidgetsController_getWidgetConfig",
        "parameters": [
          {
            "name": "widget_tag",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "app_id",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/"
                  }
                }
              }
            }
          }
        },
        "summary": "get a widget config",
        "tags": [
          "Widgets"
        ]
      }
    },
    "/api/v1/app/widgets/create": {
      "post": {
        "operationId": "WidgetsController_createWidget",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWidgetDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDto"
                }
              }
            }
          }
        },
        "summary": "create a widget",
        "tags": [
          "Widgets"
        ]
      }
    },
    "/api/v1/app/widgets/getAccessToken": {
      "post": {
        "operationId": "WidgetsController_getAccessToken",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetAccessTokenDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetAccessTokenResponseDto"
                }
              }
            }
          }
        },
        "summary": "get access token for a widget",
        "tags": [
          "Widgets"
        ]
      }
    },
    "/api/v1/app/widgets/update": {
      "post": {
        "operationId": "WidgetsController_updateWidget",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWidgetDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDto"
                }
              }
            }
          }
        },
        "summary": "update a widget",
        "tags": [
          "Widgets"
        ]
      }
    },
    "/api/v1/app/widgets/delete": {
      "post": {
        "operationId": "WidgetsController_deleteWidget",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteWidgetDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDto"
                }
              }
            }
          }
        },
        "summary": "delete a widget",
        "tags": [
          "Widgets"
        ]
      }
    },
    "/api/v1/app/widgets/applyConfigToApps": {
      "post": {
        "operationId": "WidgetsController_applyWidgetConfigToApps",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplyWidgetConfigToAppsDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDto"
                }
              }
            }
          }
        },
        "summary": "apply a widget config to apps",
        "tags": [
          "Widgets"
        ]
      }
    },
    "/api/v1/app/widgets/unbindConfigFromApps": {
      "post": {
        "operationId": "WidgetsController_unbindWidgetConfigFromApps",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnbindWidgetConfigFromAppsDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "unbind a widget config from apps",
        "tags": [
          "Widgets"
        ]
      }
    },
    "/api/v1/app/widgets/{tag}": {
      "get": {
        "operationId": "WidgetsController_getWidgetByTag",
        "parameters": [
          {
            "name": "tag",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDto"
                }
              }
            }
          }
        },
        "summary": "get a widget by tag",
        "tags": [
          "Widgets"
        ]
      }
    },
    "/api/v1/app/widgets/subscribe": {
      "post": {
        "operationId": "WidgetsController_subscribeWidget",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubscribeWidgetDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDto"
                }
              }
            }
          }
        },
        "summary": "subscribe a widget",
        "tags": [
          "Widgets"
        ]
      }
    },
    "/api/v1/app/widgets/unsubscribe": {
      "post": {
        "operationId": "WidgetsController_unsubscribeWidget",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnsubscribeWidgetDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDto"
                }
              }
            }
          }
        },
        "summary": "unsubscribe a widget",
        "tags": [
          "Widgets"
        ]
      }
    }
  },
  "info": {
    "title": "3 Body Labs API Reference",
    "description": "This is the API reference for 3 Body Labs",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "in": "header",
        "name": "Authorization",
        "description": "JWT Authorization"
      }
    },
    "schemas": {
...
      "LoginResponseDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "Token"
          }
        },
        "required": [
          "token"
        ]
      },
...
      "WidgetSettingsDto": {
        "type": "object",
        "properties": {
          "widget_tag": {
            "type": "string",
            "description": "widget tag"
          },
          "management_url": {
            "type": "string",
            "description": "management url"
          },
          "widget_url": {
            "type": "string",
            "description": "widget url"
          },
          "metadata": {
            "type": "object",
            "description": "metadata"
          },
          "permissions": {
            "type": "string",
            "description": "permissions",
            "enum": [
              "read_ip",
              "manage_ip",
              "read_portal",
              "manage_portal",
              "get_user_info",
              "manage_wallet"
            ]
          },
          "type": {
            "type": "string",
            "description": "type",
            "enum": [
              "iframe",
              "system"
            ]
          }
        },
        "required": [
          "widget_tag",
          "management_url",
          "widget_url",
          "metadata",
          "permissions",
          "type"
        ]
      },
      "WidgetSummaryDto": {
        "type": "object",
        "properties": {
          "tag": {
            "type": "string",
            "description": "widget tag"
          },
          "name": {
            "type": "string",
            "description": "widget name"
          },
          "summary": {
            "type": "string",
            "description": "widget summary"
          },
          "pricing": {
            "type": "object",
            "description": "widget pricing"
          },
          "settings": {
            "description": "widget settings",
            "allOf": [
              {
                "$ref": "#/components/schemas/WidgetSettingsDto"
              }
            ]
          },
          "is_featured": {
            "type": "boolean",
            "description": "widget is featured"
          },
          "is_new": {
            "type": "boolean",
            "description": "widget is new"
          },
          "is_official": {
            "type": "boolean",
            "description": "widget is official"
          },
          "category": {
            "type": "string",
            "description": "widget category"
          },
          "author": {
            "type": "string",
            "description": "widget author"
          },
          "icon": {
            "type": "string",
            "description": "widget icon"
          },
          "description": {
            "type": "string",
            "description": "widget description"
          },
          "coming_soon": {
            "type": "boolean",
            "description": "widget coming soon"
          },
          "priority": {
            "type": "number",
            "description": "widget priority"
          },
          "subscribers": {
            "type": "number",
            "description": "widget subscribers"
          },
          "is_subscribed": {
            "type": "boolean",
            "description": "widget is subscribed"
          }
        },
        "required": [
          "tag",
          "name",
          "settings",
          "is_featured",
          "is_new",
          "is_official",
          "category",
          "author",
          "icon",
          "description",
          "coming_soon",
          "priority",
          "subscribers",
          "is_subscribed"
        ]
      },
      "CreateWidgetDto": {
        "type": "object",
        "properties": {
          "tag": {
            "type": "string",
            "description": "widget tag"
          },
          "name": {
            "type": "string",
            "description": "widget name"
          },
          "summary": {
            "type": "string",
            "description": "widget summary"
          },
          "pricing": {
            "type": "object",
            "description": "widget pricing"
          },
          "settings": {
            "description": "widget settings",
            "allOf": [
              {
                "$ref": "#/components/schemas/WidgetSettingsDto"
              }
            ]
          },
          "is_featured": {
            "type": "boolean",
            "description": "widget is featured"
          },
          "is_new": {
            "type": "boolean",
            "description": "widget is new"
          },
          "is_official": {
            "type": "boolean",
            "description": "widget is official"
          },
          "category": {
            "type": "string",
            "description": "widget category"
          },
          "author": {
            "type": "string",
            "description": "widget author"
          },
          "icon": {
            "type": "string",
            "description": "widget icon"
          },
          "description": {
            "type": "string",
            "description": "widget description"
          },
          "coming_soon": {
            "type": "boolean",
            "description": "widget coming soon"
          },
          "priority": {
            "type": "number",
            "description": "widget priority"
          }
        },
        "required": [
          "tag",
          "name",
          "settings",
          "is_featured",
          "is_new",
          "is_official",
          "category",
          "author",
          "icon",
          "description",
          "coming_soon",
          "priority"
        ]
      },
      "GetAccessTokenDto": {
        "type": "object",
        "properties": {
          "tag": {
            "type": "string",
            "description": "widget tag"
          }
        },
        "required": [
          "tag"
        ]
      },
      "GetAccessTokenResponseDto": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string",
            "description": "access token"
          }
        },
        "required": [
          "access_token"
        ]
      },
      "UpdateWidgetDto": {
        "type": "object",
        "properties": {
          "tag": {
            "type": "string",
            "description": "widget tag"
          },
          "name": {
            "type": "string",
            "description": "widget name"
          },
          "summary": {
            "type": "string",
            "description": "widget summary"
          },
          "pricing": {
            "type": "object",
            "description": "widget pricing"
          },
          "settings": {
            "description": "widget settings",
            "allOf": [
              {
                "$ref": "#/components/schemas/WidgetSettingsDto"
              }
            ]
          },
          "is_featured": {
            "type": "boolean",
            "description": "widget is featured"
          },
          "is_new": {
            "type": "boolean",
            "description": "widget is new"
          },
          "is_official": {
            "type": "boolean",
            "description": "widget is official"
          },
          "category": {
            "type": "string",
            "description": "widget category"
          },
          "author": {
            "type": "string",
            "description": "widget author"
          },
          "icon": {
            "type": "string",
            "description": "widget icon"
          },
          "description": {
            "type": "string",
            "description": "widget description"
          },
          "coming_soon": {
            "type": "boolean",
            "description": "widget coming soon"
          },
          "priority": {
            "type": "number",
            "description": "widget priority"
          }
        }
      },
      "DeleteWidgetDto": {
        "type": "object",
        "properties": {
          "tag": {
            "type": "string",
            "description": "widget tag"
          }
        },
        "required": [
          "tag"
        ]
      },
      "ApplyWidgetConfigToAppsDto": {
        "type": "object",
        "properties": {
          "public": {
            "type": "object",
            "description": "public configuration for widget"
          },
          "private": {
            "type": "object",
            "description": "private configuration for widget"
          },
          "tag": {
            "type": "string",
            "description": "widget tag"
          },
          "app_id": {
            "type": "string",
            "description": "app id"
          },
          "enabled": {
            "type": "boolean",
            "description": "enabled"
          }
        },
        "required": [
          "public",
          "private",
          "tag",
          "app_id"
        ]
      },
      "UnbindWidgetConfigFromAppsDto": {
        "type": "object",
        "properties": {
          "tag": {
            "type": "string",
            "description": "widget tag"
          },
          "app_id": {
            "type": "string",
            "description": "app id"
          }
        },
        "required": [
          "tag",
          "app_id"
        ]
      },
      "SubscribeWidgetDto": {
        "type": "object",
        "properties": {
          "tag": {
            "type": "string",
            "description": "widget tag"
          },
          "publicConfig": {
            "type": "object",
            "description": "public configuration for widget"
          },
          "privateConfig": {
            "type": "object",
            "description": "private configuration for widget"
          }
        },
        "required": [
          "tag"
        ]
      },
      "UnsubscribeWidgetDto": {
        "type": "object",
        "properties": {
          "tag": {
            "type": "string",
            "description": "widget tag"
          }
        },
        "required": [
          "tag"
        ]
      }
    }
  },
...
}